Add a mock collector template
This template requires a parameter for the source UID you want it
to relate the new objects to.

You can also change which built-in config it uses by changing the
CONFIG parameter

If you want to specify a new config, you can overwrite the existing
one in /opt/mock-collector/config/openshift with a configmap of
your choice.