Add build config templates and make deployment configs templates
The build configs make it easier to rebuild images by not requiring
them to be manually pushed to the integrated registry.

By making both the build configs and deployment configs templates
we allow the namespace from which we pull the image to be configured
using a template parameter.

In production this namespace will be "buildfactory", but if QE or
dev wants to push a test image they don't have to polute the
buildfactory repo or their git tree.