Rename the collector image to topological-inventory-openshift
The collector and operations workers now share a repo and they can
also share an image.
The entrypoint by default will be the collector bin script so that
we don't have to encode the entrypoint for each collector into the
orchestrator, but we can override that in the operations worker
deployment config.